+++ CHANGELOG.md
@@ -26,7 +26,7 @@ Released June 26th, 2023.
 ### Changed
 
 - Vertex agent now attempts to retry create custom job up to three times to recover from transient errors - [#192](https://github.com/PrefectHQ/prefect-gcp/192)
-- Updated `prefect.docker` import fo `prefect.utilities.dockerutils` - [#194](https://github.com/PrefectHQ/prefect-gcp/pull/194)
+- Updated `prefect.docker` import to `prefect.utilities.dockerutils` - [#194](https://github.com/PrefectHQ/prefect-gcp/pull/194)
 
 ## 0.4.3